# release-20.1: opt: reduce allocations in optbuilder